Heavily refactor event creation handler
